requirements.txt
PySide6>=6.5.0
qrcode[pil]>=7.4
Pillow>=9.5.0
reportlab>=4.0
validators>=0.20.0